+++ src/utils/misc.h
@@ -27,8 +27,8 @@ namespace Utils {
         QString prependIfDoesNotStartWith(QString text, QString startString);
         QString appendIfDoesNotEndWith(QString text, QString endString);
         bool startDetachedProcess(QString executablePath,
-                                          QStringList parameters = QStringList(),
-                                          QString workingDirectory = "");
+                                  QStringList parameters = QStringList(),
+                                  QString workingDirectory = "");
         QString shorten(QString text, int length, QString sequence = "...");
         QString defaultNotesPath();
         QString dirSeparator();
@@ -44,5 +44,6 @@ namespace Utils {
         QList<QObject *> getParents(QObject *object);
         QString appDataPath();
         QString logFilePath();
+        QString transformLinefeeds(QString text);
     }
 }